Add OpenAI vision service and fallback mechanism; update cookbook and recipe models

+++ src/DigitalMealPlanner.Web/Modules/Recipes/Models/RecipeDraft.cs
@@ -10,10 +10,10 @@ public class RecipeDraft
     [JsonPropertyName("prep_minutes")] public int PrepMinutes { get; set; }
     [JsonPropertyName("cook_minutes")] public int CookMinutes { get; set; }
     [JsonPropertyName("cuisine_type")] public string CuisineType { get; set; } = string.Empty;
-    [JsonPropertyName("calories_per_serving")] public int CaloriesPerServing { get; set; }
-    [JsonPropertyName("protein_g")]    public int ProteinG { get; set; }
-    [JsonPropertyName("carbs_g")]      public int CarbsG { get; set; }
-    [JsonPropertyName("fat_g")]        public int FatG { get; set; }
+    [JsonPropertyName("calories_per_serving")] public double CaloriesPerServing { get; set; }
+    [JsonPropertyName("protein_g")]    public double ProteinG { get; set; }
+    [JsonPropertyName("carbs_g")]      public double CarbsG { get; set; }
+    [JsonPropertyName("fat_g")]        public double FatG { get; set; }
     [JsonPropertyName("tags")]         public List<string> Tags { get; set; } = [];
     [JsonPropertyName("steps")]        public List<string> Steps { get; set; } = [];
     [JsonPropertyName("ingredients")]  public List<IngredientDraft> Ingredients { get; set; } = [];
@@ -25,5 +25,5 @@ public class IngredientDraft
     [JsonPropertyName("quantity")]          public string Quantity { get; set; } = string.Empty;
     [JsonPropertyName("unit")]              public string Unit { get; set; } = string.Empty;
     [JsonPropertyName("category")]          public string Category { get; set; } = string.Empty;
-    [JsonPropertyName("calories_per_unit")] public int CaloriesPerUnit { get; set; }
+    [JsonPropertyName("calories_per_unit")] public double CaloriesPerUnit { get; set; }
 }
